Implement form functionality for Header component
The header component now holds all the user data in its own state
object. The state object also has a editMode flag that is set to true
when the user clicks the Edit button.

If the editMode flag is set to true, the Header component will render
input fields in the place of all the elements that display the user
information within the header.