Implement the Experience component
The experience component will have the functionality to add more
experience entries when the user submits the add experience form.

Individual experiences are able to be deleted.

This component still needs to allow the user to add a bullet point list
of accomplishments for each experience entry. This will likely be done
by adding an array property within each experience object that holds
a list of user-input strings that contain the description of
accomplishments.